requirements.txt
instagrapi==2.1.3
python-dotenv==1.0.1

Pillow>=8.1.1

# annotated-types==0.7.0
# certifi==2024.8.30
# charset-normalizer==3.4.0
# idna==3.10
# pycryptodomex==3.21.0
# pydantic==2.10.1
# pydantic_core==2.27.1
# PySocks==1.7.1
# requests==2.32.3
# typing_extensions==4.12.2
# urllib3==2.2.3
